Update sort.sh
Add step with macs2 peak calling

+++ ATACseq_footprinting/sort.sh
@@ -81,22 +81,30 @@ else
       echo "sorted merged bams already indexed"
 fi      
 
-##############################---STEP 6: INDEX MERGED BAMS---###########################################
+###########################---STEP 6: PEAK CALLING WITH MACS2---########################################
+if [ ! -f "$bam_path/peak_calling/${PREFIX}/${PREFIX}_raw.bed" ]; then
 
-gsize=2620345972 #is this correct?!
-macs="$bam_path/peak_calling/${PREFIX}/${PREFIX}_peaks.broadPeak"
-raw="$bam_path/peak_calling/${PREFIX}/${PREFIX}_raw.bed"
-log="$bam_path/Logs/${PREFIX}_peak_calling.log"
+      if [ ! -f "$bam_path/peak_calling" ]; then
+            mkdir "$bam_path/peak_calling/"
+      fi
+      if [ ! -f "$bam_path/peak_calling/${PREFIX}" ]; then
+            mkdir "$bam_path/peak_calling/${PREFIX}"
+      fi
+      
+      module load macs2 
+      gsize=2620345972 #is this correct?!
+      echo "Running macs2 with .bam file: ${PREFIX}.merged.sorted.bam"
 
-echo "Running macs2 with .bam file: ${PREFIX}.merged.sorted.bam"
+      #Arguments:
+      #-t: the IP data file
+      #--outdir: the output directory
 
-#Arguments:
-#-t: the IP data file
-#--outdir: the output directory
-
-mkdir "$bam_path/peak_calling/${PREFIX}"
-
-macs2 callpeak -t "$bam_path/${PREFIX}.merged.sorted.bam" --name ${PREFIX} --outdir "$bam_path/peak_calling/${PREFIX}" --gsize $gsize --nomodel --shift -100 --extsize 200 --broad &> "$bam_path/Logs/{PREFIX}_peak_calling.log"
-cp "$bam_path/peak_calling/${PREFIX}/{PREFIX}_peaks.broadPeak" "$bam_path/peak_calling/${PREFIX}/{PREFIX}_raw.bed"
-
+      macs2 callpeak -t "$bam_path/${PREFIX}.merged.sorted.bam" --name ${PREFIX} --outdir "$bam_path/peak_calling/${PREFIX}" --gsize $gsize --nomodel --shift -100 --extsize 200 --broad 
+      cp "$bam_path/peak_calling/${PREFIX}/{PREFIX}_peaks.broadPeak" "$bam_path/peak_calling/${PREFIX}/{PREFIX}_raw.bed"
+     
+      echo "peak calling with macs2 complete"
+else
+      echo "peaking calling with macs2 already done"
+fi
+
